feature: improved tagging ui for single media items feature: long press to delete tag

+++ app/src/main/java/com/kaii/photos/repositories/TaggedItemsRepository.kt
@@ -7,21 +7,25 @@ import com.kaii.photos.database.entities.TaggedItem
 class TaggedItemsRepository(
     private val dao: TaggedItemsDao
 ) {
-    suspend fun addTag(tag: Tag, mediaId: Long) =
-        dao.upsert(
-            TaggedItem(
-                tag = tag.id,
-                mediaId = mediaId
+    suspend fun addTag(tag: Tag, mediaId: Long) {
+        if (mediaId > 0L) {
+            dao.upsert(
+                TaggedItem(
+                    tag = tag.id,
+                    mediaId = mediaId
+                )
             )
-        )
+        }
+    }
 
-    suspend fun removeTag(tag: Tag, mediaId: Long) =
-        dao.remove(
-            TaggedItem(
-                tag = tag.id,
-                mediaId = mediaId
+    suspend fun removeTag(tag: Tag, mediaId: Long) {
+        if (mediaId > 0L) {
+            dao.remove(
+                TaggedItem(
+                    tag = tag.id,
+                    mediaId = mediaId
+                )
             )
-        )
-
-    suspend fun getItem(mediaId: Long) = dao.getItem(id = mediaId)
+        }
+    }
 }
